De-nest models directory
