+++ app/faceanalysis/face_matcher.py
@@ -10,12 +10,12 @@ from faceanalysis.face_vectorizer import face_vector_from_text, FaceVector
 from faceanalysis.face_vectorizer import face_vector_to_text
 from faceanalysis.face_vectorizer import get_face_vectors
 from faceanalysis.log import get_logger
-from faceanalysis.models.image_status_enum import ImageStatusEnum
-from faceanalysis.models.models import FeatureMapping
-from faceanalysis.models.models import Image
-from faceanalysis.models.models import ImageStatus
-from faceanalysis.models.models import Match
-from faceanalysis.models.models import get_db_session
+from faceanalysis.models import FeatureMapping
+from faceanalysis.models import Image
+from faceanalysis.models import ImageStatus
+from faceanalysis.models import ImageStatusEnum
+from faceanalysis.models import Match
+from faceanalysis.models import get_db_session
 from faceanalysis.settings import DISTANCE_SCORE_THRESHOLD
 from faceanalysis.settings import FACE_VECTORIZE_ALGORITHM
 from faceanalysis.storage import StorageError
