--- app/faceanalysis/models/image_status_enum.py
@@ -1,8 +1,0 @@
-from enum import Enum
-
-
-class ImageStatusEnum(Enum):
-    finished_processing = 4
-    processing = 3
-    uploaded = 1
-    face_vector_computed = 5
